fix: add missing merge changes
