+++ apps/main-landing/src/app/vault/hooks/use-staking.ts
@@ -9,12 +9,12 @@ import {
   useWriteContract,
 } from 'wagmi';
 import { useConnectModal } from '@rainbow-me/rainbowkit';
+import { STAKER_ADDRESS, StakingABI } from '@idriss-xyz/constants';
 
 import { useGetStakedBalance } from '@/app/vault/commands/get-staked-balance';
 import { useGetBonusStakedBalance } from '@/app/vault/commands/get-bonus-staked-balance';
 import { useGetUnstakedBalance } from '@/app/vault/commands/get-unstaked-balance';
 import { formatNumber } from '@/app/claim/components/claim/components/idriss-user-criteria-description';
-import { STAKER_ADDRESS, StakingABI } from '@/app/vault/constants';
 import { ERC20_ABI } from '@/app/creators/donate/constants';
 import { IDRISS_TOKEN_ADDRESS } from '@/components/token-section/constants';
